broke volunteerSlotRoster into its own component -- currently gets all Volunteer Slots
