+++ src/components/admin/VolunteerSlotPost.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export const VolunteerSlotPost = () => {
+
+  return(
+    
+  )
+}
